Changed search function to handleSubmit
I prefer the term handleSubmit as a function name to handle the user's submission in a form. Will divide it into smaller functions, like in the solution. Can then use it to call the next function. Nice to separate a function out rather than bunch it all together, as I did.