Activated 'Current Location' Button
I added the variable, event listener and two functions needed to acquire information about current weather conditions of current location when user clicks on the button 'Current Location'. The first function, getCurrentLocation, finds what the current geolocation coordinates are. The second function, searchLocation includes the api Key, URL and axios to find the weather data of current coordinates. This function THEN calls the showWeather function which displays the wanted information in the inner HTML for the user to see.