Fixed: Current location button to display forecast
I added the api key, api URL for the forecast the axios for the forecast URL into the showWeather function, and called the displayWeather function from inside the showWeather function. This means that now, clicking on the current location button will ALSO display the desired forecast for the entered city.